Initial working Universal React Version
