+++ package.json
@@ -46,13 +46,13 @@
     "react-dom": "^15.4.2",
     "react-render-html": "^0.1.6",
     "react-router": "^3.0.2",
+    "react-router-bootstrap": "^0.23.1",
     "webpack-dev-server": "^2.3.0"
   },
   "scripts": {
-    "start": "concurrently \"npm run server\" \"npm run client\"",
-    "startwpdev": "concurrently \"NODE_ENV=production node_modules/.bin/babel-node --presets react,es2015 src/server.js\" \"node_modules/.bin/http-server src/static\"",
-    "server": "babel-node ./src/server.js",
-    "client": "babel-node ./src/app-client.js",
+    "start": "NODE_ENV=production node_modules/.bin/webpack -p && concurrently \"npm run server\" \"npm run client\"",
+    "server": "NODE_ENV=production node_modules/.bin/babel-node --presets react,es2015 src/server.js",
+    "client": "node_modules/.bin/http-server src/static",
     "dev": "echo \"This command has been deprecated. Use 'npm start'\" && exit 1"
   },
   "devDependencies": {
