+++ src/app-client.js
@@ -1,7 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import AppRoutes from './components/AppRoutes';
+import AppRoutes from './components/AppRoutes.jsx';
 
-window.onload = () => {
-  ReactDOM.render(<AppRoutes/>, document.getElementById('app'));
-};
+ReactDOM.render(<AppRoutes/>, document.getElementById('app'));
