--- src/client-render.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { Router } from 'react-router';
-//import { routes } from './routes';
-import AppRoutes from './components/AppRoutes';
-
-
-
-ReactDOM.render(<AppRoutes />, document.getElementById('app'));
