+++ src/components/Footer.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 class Footer extends React.Component {
   render() {
     return(
-      <div>
+      <div className="site-footer">
         <p>© 2017 All Rights Reserved.</p>
 
         <p>Legal Disclaimer: The site DOES NOT provide any legal advice, only information. Users of this web site should consult with their own lawyer for legal advice.</p>
