+++ src/components/Home.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import Header from './Header';
-import Footer from './Footer';
+import Header from './Header.jsx';
+import Footer from './Footer.jsx';
 import { Col, Row } from 'react-bootstrap';
 
 class Home extends React.Component {
