--- src/static/index.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no">
-    <title>NC Reentry Resources Hub</title>
-
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/latest/css/bootstrap.min.css" />
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/latest/css/bootstrap-theme.min.css" />
-    <!-- <link rel="stylesheet" href="/styles/style.css"> -->
-  </head>
-  <body>
-    <div id="app"></div>
-    <script src="/scripts/bundle.js"></script>
-  </body>
-</html>
