+++ src/static/styles/style.css
@@ -1,0 +1,37 @@
+body {
+  font-family: 'PT Serif', serif;
+}
+
+.site-header {
+  font-family: Lato, sans-serif;
+}
+
+.header-background-image {
+  background-image: url(http://dev-big-old-book.pantheonsite.io/wp-content/uploads/2016/09/cropped-View_looking_north_of_Mount_Jefferson-1.jpg);
+  background-repeat: no-repeat;
+  background-size: cover;
+  background-position: center bottom;
+  opacity: 0.75;
+  filter: alpha(opacity=75); /* For IE8 and earlier */
+  background-attachment: fixed;
+  padding: 4rem 0;
+}
+
+.title-box {
+  text-align: center;
+  color: #fff;
+  width: 750px;
+  max-width: 75%;
+  padding: 1rem;
+  margin: 0 auto;
+  background-color: hsla(198, 95%, 38%, 0.75);
+}
+
+.site-footer {
+  font-family: Lato, sans-serif;
+  padding: 2em;
+  font-size: 1.6rem;
+  color: #fff;
+  background: hsl(198, 95%, 38%);
+  text-align: center;
+}
