+++ webpack.config.js
@@ -10,7 +10,7 @@ module.exports = {
   module: {
     loaders: [
       {
-        test: /\.js$/,
+        test: /.jsx?$/,
         loader: 'babel-loader',
         exclude: /node_modules/,
         query: {
@@ -33,8 +33,8 @@ module.exports = {
       compress: { warnings: false },
       mangle: false,
       sourcemap: true,
-      beautify: true,
-      dead_code: true
+      beautify: false,
+      dead_code: false
     })
   ]
 };
